fix: H2-console 접근 불가 버그 수정
