+++ src/main/java/com/zero/bwtableback/configuration/SecurityConfig.java
@@ -1,0 +1,46 @@
+package com.zero.bwtableback.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+
+        http
+                .csrf(csrf -> csrf.disable()) // CSRF 비활성화
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("/h2-console/**").permitAll() // H2 콘솔 접근 허용
+                        .anyRequest().authenticated() // 나머지 요청은 인증 필요
+                )
+                .headers(headers -> headers
+                        .frameOptions(frameOptions -> frameOptions.disable()) // H2 콘솔을 iframe에서 사용할 수 있도록 설정
+                );
+
+//    @Bean
+//    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+//        http
+//                .authorizeRequests()
+//                .anyRequest().authenticated()
+//                .and()
+//                .oauth2Login(oauth2 -> oauth2
+//                        .userInfoEndpoint(userInfo -> userInfo
+//                                .userService(principalOauth2UserService)
+//                        )
+//                );
+
+        //    private final PrincipalOauth2UserService principalOauth2UserService;
+//
+//    public SecurityConfig(PrincipalOauth2UserService principalOauth2UserService) {
+//        this.principalOauth2UserService = principalOauth2UserService;
+//    }
+
+        return http.build();
+    }
+}
